moved help files to web-src/help. put all help content in containers and added header.
