+++ Internal Data Source/web-src/help/dpHelp.html
@@ -2,7 +2,8 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
-<h1>Overview</h1>
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dox.internalPP"></h1>
 <p>
   Internal data points can be created to monitor specific values that monitor the internal workings of Mango Automation.
 </p>
@@ -121,3 +122,4 @@ This value represents the maximum amount of memory that the heap can potentially
 <p>
 This value represents the number of CPU cores that the JVM can access.
 </p>
+</div>
