+++ Internal Data Source/web-src/help/dsHelp.html
@@ -2,7 +2,8 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
-<h1>Overview</h1>
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dox.internalDS"></h1>
 <p>
   The Internal data source is used to monitor various performance aspects of Mango Automation. 
 </p>
@@ -12,3 +13,4 @@
   create points for new monitor IDs that match the <b>Create points pattern</b>. If no pattern is supplied, no points
   are automatically created. If you wish to use more than one of the supplied patterns, you may conjoin them as
 </p>
+</div>
