+++ Internal Data Source/web-src/internal.js
@@ -6,8 +6,8 @@
 import angular from 'angular';
 import internalDataSourceEditor from './components/InternalDataSourceEditor/InternalDataSourceEditor';
 import internalDataPointEditor from './components/InternalDataPointEditor/InternalDataPointEditor';
-import dsHelpTemplate from './dsHelp.html';
-import dpHelpTemplate from './dpHelp.html';
+import dsHelpTemplate from './help/dsHelp.html';
+import dpHelpTemplate from './help/dpHelp.html';
 
 const internalDataSourceModule = angular.module('maInternalDataSource', ['maUiApp'])
 .component('maInternalDataSourceEditor', internalDataSourceEditor)
