+++ Log4J reset/web-src/help/helpPage.html
@@ -1,3 +1,4 @@
+<div class="ma-help-doc-container">
 <h1 ma-tr="log4JReset.settings.header"></h1>
 
 <h2></h2>
@@ -31,3 +32,4 @@
     </Loggers>
     </pre>
 </p>
+</div>
