+++ Maintenance Events/web-src/help/helpPage.html
@@ -1,4 +1,5 @@
-<h2>Overview</h2>
+<div class="ma-help-doc-container">
+    <h1 ma-tr="eventHandlers.maintenanceEvents"></h1>
 
 <p>
     Maintenance events are used to schedule maintenance periods for data sources, such that 
@@ -71,3 +72,4 @@
 <p>
     To manually enable or disable a maintenance event, click the <b>Toggle Event</b> switch at the top right.
 </p>
+</div>
