+++ Mango API/web-src/help/helpPage.html
@@ -1,4 +1,5 @@
-<h2>Swagger UI</h2>
+<div class="ma-help-doc-container">
+    <h1 ma-tr="mangoapi.description"></h1>
 <p>
 	Mango's REST uses Swagger to allow code auto-generation and includes an API exploration tool at /swagger-ui.html.  
 	Note that this is disabled by default.  To enable this feature set the env property swagger.enabled=true.  
@@ -21,3 +22,4 @@
      To enable the filter you must edit the <b>rest.cors.*</b> section in the <b>env.properties</b> file and 
      restart Mango. Note that changes to this file are only picked up upon a restart of Mango.
 </p>
+</div>
