+++ VMStat data source/web-src/help/dpHelp.html
@@ -2,9 +2,11 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
-<h1>Overview</h1>
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dox.vmstatPP"></h1>
 <p>
   VMStat points can be created to monitor specific values that are available in vmstat output. For more detail on the 
   available attributes, see the <a href="http://linuxcommand.org/man_pages/vmstat8.html" target="_blank">linux 
   documentation</a>.
 </p>
+</div>
