+++ VMStat data source/web-src/help/dsHelp.html
@@ -2,7 +2,8 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
-<h1>Overview</h1>
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dox.vmstatDS"></h1>
 <p>
   The VMStat data source is used to monitor various performance aspects of a *nix host. When the data source starts a
   process is initiated that runs the vmstat program, and the output from that process is monitored.
@@ -14,3 +15,4 @@
   how often the vmstat process will report data back to Mango. The <b>Output scale</b> can be set to automatically 
   adjust byte values to different scales, including bytes, k, KB, m, and MB.
 </p>
+</div>
