+++ Virtual Data Source/web-src/help/dpHelp.html
@@ -2,7 +2,8 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
-<h1>Overview</h1>
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dsEdit.virtualPoint"></h1>
 <p>Points within a Virtual Data Source generate data without any external influence, either randomly or deterministically. Any supported data type can be created, and a number of options for changing values are available.</p>
 
 <p>
@@ -45,3 +46,4 @@
 
 <h1>Alphanumeric Points</h1>
 <p>Alphanumeric points represent strings of characters, and thus are not subject to meaningful change via algorithm. As such, the only available change type is <b>No change</b>.</p>
+</div>
