+++ Virtual Data Source/web-src/help/dsHelp.html
@@ -2,6 +2,8 @@
     Copyright (C) 2014 Infinite Automation Systems Inc. All rights reserved.
     @author Matthew Lohbihler
 -->
+<div class="ma-help-doc-container">
+  <h1 ma-tr="dsEdit.virtual"></h1>
 <p>The Virtual Data Source is a data source without any connection to an external or physical system. It exists entirely within the system's own memory space. The data it produces is either random, or deterministic depending on how you configure the points.</p>
 
 <p>
@@ -19,3 +21,4 @@
   If one unchecks the <b>Perform polls</b> checkbox, the virtual data source will not generate updates for the points. This can be useful if you are using virtual points as components in more complex control logic, and only wish for the points to update when
   they are set from a point link, a script, the interface, etc.
 </p>
+</div>
